src/test/java/guru/springframework/didemo/controllers/PropertyInjectedControllerTest.java
/**
 * 
 */
package guru.springframework.didemo.controllers;

import org.junit.Before;
import org.junit.Test;

import guru.springframework.didemo.services.GreetingServiceImpl;

import static org.junit.Assert.assertEquals;

/**
 * Very bad way to do this!!! <br>
 * 
 * @author IzaakVanNiekerk
 *
 */
public class PropertyInjectedControllerTest {

	private PropertyInjectedController propertyInjectedController;

	@Before
	public void setUp() throws Exception {
		this.propertyInjectedController = new PropertyInjectedController();
		this.propertyInjectedController.greetingService = new GreetingServiceImpl();
	}

	@Test
	public void testGreeting() throws Exception {
		assertEquals(GreetingServiceImpl.HELLO_GURUS, propertyInjectedController.sayHello());
	}

}
